Add files via upload
This webpage generates a random number when the page is reloaded. The user has to guess the number. When the user enters the number in the input box and clicks on the check button, Java script code compares the guessed number with already generated random number and prints respective message on the screen that helps the user to guess the number.